Use alternative method to delete from list following exception in Github module
When trying to remove an item from the list and the item contains a GitHubPullRequestReview object
then it will raise an exception in github.GithubObject.CompletableGithubObject#__eq__ due to no _url_ property.

This MP uses an alternative method of remove an item from a list, specifically using the index rather than
comparing items in the list.