Use the name of the person opening a PR as author.
This is more consistent with the github UI.

In Exoscale's case, the workflow is that developers create branches in
shared repos (as opposed to forking repositories). This change should
therefore make the owner column useful, while organisations that use the
forking workflow should see no difference (the people opening the PRs
are very very often the owner of the fork).