Use the name of the person opening a PR as author. (#23)
This is more consistent with the github UI.

In organizations which use a shared repo (as opposed to forking
repositories), the repo owner is common. This change should
therefore make the owner column useful, while organisations
that use the forking workflow should see no difference (the people
opening the PRs are very very often the owner of the fork).